tests - remaining tests refactor
