+++ tests/Backend/Common/TableExporterTest.php
@@ -7,11 +7,14 @@
  * To change this template use File | Settings | File Templates.
  */
 
+
+namespace Keboola\Test\Backend\Common;
+use Keboola\Test\StorageApiTestCase;
 use Keboola\StorageApi\Client,
 	Keboola\Csv\CsvFile,
 	Keboola\StorageApi\TableExporter;
 
-class Keboola_StorageApi_Tables_TableExporterTest extends StorageApiTestCase
+class TableExporterTest extends StorageApiTestCase
 {
 
 
@@ -21,16 +24,16 @@ class Keboola_StorageApi_Tables_TableExporterTest extends StorageApiTestCase
 	public function setUp()
 	{
 		parent::setUp();
-		$this->_initEmptyBucketsForAllBackends();
-		$this->downloadPath = __DIR__ . '/../_tmp/languages.sliced.csv';
-		$this->downloadPathGZip = __DIR__ . '/../_tmp/languages.sliced.csv.gz';
+		$this->_initEmptyTestBuckets();
+		$this->downloadPath = __DIR__ . '/../../_tmp/languages.sliced.csv';
+		$this->downloadPathGZip = __DIR__ . '/../../_tmp/languages.sliced.csv.gz';
 	}
 
 	/**
 	 * @dataProvider tableImportData
 	 * @param $importFileName
 	 */
-	public function testTableAsyncExportRepeatedly($backend, CsvFile $importFile, $expectationsFileName, $exportOptions=array())
+	public function testTableAsyncExportRepeatedly(array $supportedBackends, CsvFile $importFile, $expectationsFileName, $exportOptions=array())
 	{
 		$oldRunId = $this->_client->generateRunId();
 		$this->_client->setRunId($oldRunId);
@@ -39,7 +42,7 @@ class Keboola_StorageApi_Tables_TableExporterTest extends StorageApiTestCase
 			$exportOptions['gzip'] = false;
 		}
 
-		$tableId = $this->_client->createTable($this->getTestBucketId(self::STAGE_IN, $backend), 'languages', $importFile);
+		$tableId = $this->_client->createTable($this->getTestBucketId(self::STAGE_IN), 'languages', $importFile);
 		$result = $this->_client->writeTable($tableId, $importFile);
 
 		$this->assertEmpty($result['warnings']);
@@ -94,15 +97,19 @@ class Keboola_StorageApi_Tables_TableExporterTest extends StorageApiTestCase
 	 * @dataProvider tableImportData
 	 * @param $importFileName
 	 */
-	public function testTableAsyncExport($backend, CsvFile $importFile, $expectationsFileName, $exportOptions=array())
+	public function testTableAsyncExport(array $supportedBackends, CsvFile $importFile, $expectationsFileName, $exportOptions=array())
 	{
-		$expectationsFile = __DIR__ . '/../_data/' . $expectationsFileName;
+		$expectationsFile = __DIR__ . '/../../_data/' . $expectationsFileName;
+		$tokenData = $this->_client->verifyToken();
+		if (!in_array($tokenData['owner']['defaultBackend'], $supportedBackends)) {
+			return;
+		}
 
 		if (!isset($exportOptions['gzip']) ) {
 			$exportOptions['gzip'] = false;
 		}
 
-		$tableId = $this->_client->createTable($this->getTestBucketId(self::STAGE_IN, $backend), 'languages', $importFile);
+		$tableId = $this->_client->createTable($this->getTestBucketId(self::STAGE_IN), 'languages', $importFile);
 		$result = $this->_client->writeTable($tableId, $importFile);
 
 		$this->assertEmpty($result['warnings']);
@@ -125,17 +132,12 @@ class Keboola_StorageApi_Tables_TableExporterTest extends StorageApiTestCase
 		// compare data
 		$this->assertTrue(file_exists($this->downloadPath));
 		$this->assertLinesEqualsSorted(file_get_contents($expectationsFile), file_get_contents($this->downloadPath), 'imported data comparsion');
-
 	}
 
-	/**
-	 * @param $backend
-	 * @dataProvider backends
-	 */
-	public function testLimitParameter($backend)
+	public function testLimitParameter()
 	{
 		$importFile = new CsvFile('https://s3.amazonaws.com/keboola-tests/languages.csv');
-		$tableId = $this->_client->createTable($this->getTestBucketId(self::STAGE_IN, $backend), 'languages', $importFile);
+		$tableId = $this->_client->createTable($this->getTestBucketId(self::STAGE_IN), 'languages', $importFile);
 		$this->_client->writeTable($tableId, $importFile);
 
 		$exportOptions = array(
@@ -151,18 +153,18 @@ class Keboola_StorageApi_Tables_TableExporterTest extends StorageApiTestCase
 	public function tableImportData()
 	{
 		return array(
-			array(self::BACKEND_MYSQL, new CsvFile('https://s3.amazonaws.com/keboola-tests/languages.csv.gz'), 'languages.csv'),
-			array(self::BACKEND_MYSQL, new CsvFile('https://s3.amazonaws.com/keboola-tests/languages.csv.gz'), 'languages.csv', array('gzip' => true)),
-			array(self::BACKEND_MYSQL, new CsvFile('https://s3.amazonaws.com/keboola-tests/numbers.csv'), 'numbers.csv'),
-			array(self::BACKEND_MYSQL, new CsvFile('https://s3.amazonaws.com/keboola-tests/numbers.csv'), 'numbers.two-cols.csv', array('columns' => array('0', '45'))),
+			array([self::BACKEND_MYSQL, self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/languages.csv.gz'), 'languages.csv'),
+			array([self::BACKEND_MYSQL, self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/languages.csv.gz'), 'languages.csv', array('gzip' => true)),
+			array([self::BACKEND_MYSQL, self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/numbers.csv'), 'numbers.csv'),
+			array([self::BACKEND_MYSQL, self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/numbers.csv'), 'numbers.two-cols.csv', array('columns' => array('0', '45'))),
 
-			array(self::BACKEND_REDSHIFT, new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.backslash.redshift.out.csv', array('format' => 'escaped')),
-			array(self::BACKEND_REDSHIFT, new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.backslash.redshift.out.csv', array('format' => 'escaped')),
-			array(self::BACKEND_REDSHIFT, new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.raw.redshift.out.csv', array('format' => 'raw')),
-			array(self::BACKEND_REDSHIFT, new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.raw.redshift.out.csv', array('gzip' => true, 'format' => 'raw')),
-			array(self::BACKEND_REDSHIFT, new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.standard.out.csv', array('gzip' => true)),
-			array(self::BACKEND_REDSHIFT, new CsvFile('https://s3.amazonaws.com/keboola-tests/numbers.csv'), 'numbers.csv', array('gzip' => true)),
-			array(self::BACKEND_REDSHIFT, new CsvFile('https://s3.amazonaws.com/keboola-tests/numbers.csv'), 'numbers.two-cols.csv', array('gzip' => true, 'columns' => array('0', '45'))),
+			array([self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.backslash.redshift.out.csv', array('format' => 'escaped')),
+			array([self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.backslash.redshift.out.csv', array('format' => 'escaped')),
+			array([self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.raw.redshift.out.csv', array('format' => 'raw')),
+			array([self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.raw.redshift.out.csv', array('gzip' => true, 'format' => 'raw')),
+			array([self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/escaping.csv'), 'escaping.standard.out.csv', array('gzip' => true)),
+			array([self::BACKEND_MYSQL, self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/numbers.csv'), 'numbers.csv', array('gzip' => true)),
+			array([self::BACKEND_MYSQL, self::BACKEND_REDSHIFT], new CsvFile('https://s3.amazonaws.com/keboola-tests/numbers.csv'), 'numbers.two-cols.csv', array('gzip' => true, 'columns' => array('0', '45'))),
 		);
 	}
