--- tests/Keboola/StorageApi/_tmp/.gitignore
@@ -1,2 +1,0 @@
-*
-!.gitignore
